fix: compute the sticky refresh offset in double
CodeQL flagged cs/loss-of-precision on GetStickyRefreshDelay (#20): the
offset multiplied two ints and only then widened the result to double for
TimeSpan.FromMilliseconds, so the arithmetic could overflow int before the
conversion.

It cannot overflow in practice - the guard above pins sendNumber to 2..3,
giving 3000 or 6000 - but the rule is right that the product is computed
in the narrower type. Cast the interval to double so the multiplication
happens in double, which makes the guarantee local to the expression
instead of resting on the caller-facing guard.

Values are unchanged: the refresh delays remain exactly 3s and 6s, as
asserted by SendMessageAsync_StickyMessageRefreshesTwiceWithoutBlocking-
InitialDelivery.

+++ Messaging/MessageDeliveryPolicy.cs
@@ -18,7 +18,9 @@ internal static class MessageDeliveryPolicy
             throw new ArgumentOutOfRangeException(nameof(sendNumber));
         }
 
-        return TimeSpan.FromMilliseconds((sendNumber - 1) * StickyMessageIntervalMs);
+        // Widen before multiplying so the offset is computed in double rather than
+        // overflowing int and then being converted (cs/loss-of-precision).
+        return TimeSpan.FromMilliseconds((sendNumber - 1) * (double)StickyMessageIntervalMs);
     }
 
     internal static int GetEffectiveVisibilityDurationMs(bool useStickyMessages, int configuredTimeoutMs)
